+++ src/core/music/mediaLibrary.ts
@@ -2,6 +2,7 @@ import { buildLyricInfo } from './utils'
 import { parseLyric } from './local'
 import { invalidateCache, upsertCacheEntry } from '@/core/mediaLibrary/cache'
 import { buildMediaLibraryCacheFilePath } from '@/core/mediaLibrary/cachePath'
+import { resolveDownloadResult } from '@/core/mediaLibrary/downloadResult'
 import { resolvePlayableResource } from '@/core/mediaLibrary/playbackResolver'
 import { getMediaLibraryRuntimeRegistry } from '@/core/mediaLibrary/runtimeRegistry'
 import { mediaLibraryRepository } from '@/core/mediaLibrary/storage'
@@ -36,7 +37,9 @@ const downloadRemoteFile = async(musicInfo: LX.Music.MusicInfoRemoteFile, target
     pathOrUri: musicInfo.meta.mediaLibrary!.remotePathOrUri,
   }
   const downloadResult = await provider.downloadToCache(connection, sourceItem, targetPath)
-  if (downloadResult?.promise) await downloadResult.promise
+  await resolveDownloadResult(downloadResult, {
+    operation: `media library ${musicInfo.source} download`,
+  })
   return targetPath
 }